Various fixes and improvements whilst working on aegis.

+++ maverick-base/src/main/java/com/sshtools/common/sftp/MultipartTransfer.java
@@ -30,7 +30,7 @@ import com.sshtools.common.permissions.PermissionDeniedException;
 
 public interface MultipartTransfer {
 	
-	int getMinimumPartSize();
+	Long getMinimumPartSize();
 	
 	String getPath();
